additional tidy up
