+++ .vscode/settings.json
@@ -5,6 +5,7 @@
   },
   "files.eol": "\n",
   "cSpell.words": [
+    "agentpool",
     "alkemio",
     "mailslurper"
   ]
