+++ service/src/core/bootstrap/bootstrap.service.ts
@@ -1,16 +1,19 @@
-import pkg  from '@nestjs/common';
+import pkg from '@nestjs/common';
 const { Inject, Injectable } = pkg;
 import { ConfigService } from '@nestjs/config';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { ConfigurationTypes, LogContext } from '../../common/enums/index.js';
 import { BootstrapException } from '../../common/exceptions/bootstrap.exception.js';
 import { CommunicationAdapter } from '../../services/communication-adapter/communication.adapter.js';
+import { MatrixAdminUserElevatedService } from '@src/domain/matrix-admin/user-elevated/matrix.admin.user.elevated.service.js';
+import { MatrixAdminUserService } from '@src/domain/matrix-admin/user/matrix.admin.user.service.js';
 
 @Injectable()
 export class BootstrapService {
   constructor(
     private configService: ConfigService,
-    private communicationAdapter: CommunicationAdapter,
+    private userManagementService: MatrixAdminUserService,
+    private userElevatedManagementService: MatrixAdminUserElevatedService,
     @Inject(WINSTON_MODULE_NEST_PROVIDER)
     private readonly logger: pkg.LoggerService
   ) {}
@@ -22,9 +25,33 @@ export class BootstrapService {
         LogContext.BOOTSTRAP
       );
       this.logConfiguration();
-      await this.communicationAdapter.initiateMatrixConnection();
+      await this.initiateMatrixConnection();
     } catch (error: any) {
       throw new BootstrapException(error.message);
+    }
+  }
+
+  public async initiateMatrixConnection() {
+    await this.logServerVersion();
+    this.logger.verbose?.(
+      `Matrix admin identifier: ${this.userElevatedManagementService.adminCommunicationsID}`,
+      LogContext.BOOTSTRAP
+    );
+    await this.userElevatedManagementService.getMatrixAgentElevated();
+  }
+
+  private async logServerVersion() {
+    try {
+      const version = await this.userManagementService.getServerVersion();
+      this.logger.verbose?.(
+        `Synapse server version: ${version}`,
+        LogContext.BOOTSTRAP
+      );
+    } catch (error: any) {
+      this.logger.verbose?.(
+        `Unable to get synapse server version: ${error}`,
+        LogContext.BOOTSTRAP
+      );
     }
   }
